requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
numpy==1.21.2
pillow==8.3.1
gunicorn==20.1.0
werkzeug==2.0.3